Factor out non-nginx things into plugins
